+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.3.3] - 2021-05-03 
+- **[Added]** `SaveFlowState`, `SetFlowFromBoardState`,`SetBoardState`,`GetBoardState` and `SetFlowState`. Three states that help with board access.
+- **[Added]** Tests for the three flow access states.
+
 ## [0.3.2] - 2021-04-16
 - **[Added]** `AtLeastOneState` that ends when one of its children states completes (all are run parallely) and return `success`.